Fix some (benign) shadowing warnings

+++ src/faster/R2Stretcher.cpp
@@ -369,9 +369,9 @@ size_t
 R2Stretcher::roundUp(size_t value)
 {
     if (!(value & (value - 1))) return value;
-    int bits = 0;
+    size_t bits = 0;
     while (value) { ++bits; value >>= 1; }
-    value = 1 << bits;
+    value = size_t(1) << bits;
     return value;
 }
